requirements.dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==21.2.0             # via pytest
importlib-metadata==4.0.1  # via pluggy, pytest
iniconfig==1.1.1          # via pytest
packaging==20.9           # via pytest
pluggy==0.13.1            # via pytest
py==1.10.0                # via pytest
pyparsing==2.4.7          # via packaging
pytest==6.2.4             # via fhir_proxy (setup.py)
pytest-flask==1.0.0       # via fhir_proxy (setup.py)
pytest-mock==3.1.1        # via fhir_proxy (setup.py)
toml==0.10.2              # via pytest
typing-extensions==3.10.0.0  # via importlib-metadata
zipp==3.4.1               # via importlib-metadata
